# always use f32 for rand source of randn 